iOS: ensure keyboard is visible, even when IM hints hasn't changed
When showing a QWindow, we transfer first responder status to its
QUIView. If another QWindow was active with a text responder at
that point, the text responder will loose first responder status
in favor of the new view, and the keyboard will hide.

Now, if the new window has a focus object with the same IM state
as the previous focus object (in the previous window), m_imeState
will not change, and QIOSIntegration::update() will assume that
nothing needs to be done to show the keyboard, even if it's
actually hidden.

This patch will change the logic, so that we:
- show the keyboard if its supposed to be visible, even if
    m_imeState did not change.
- Only recreate the text responder if it needs a different
    configuration than the one we already got (not only
    from changes to Qt::ImEnabled)

Task-number: QTBUG-40695
Change-Id: I6f6788af4cbff5c7abe4f5a29e23a7cefea6b711
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@theqtcompany.com>