iOS: Check if window was active on setVisible(false) through native APIs
When a QWindow is closed and destroyed it resets QGuiApp::focus_window
before calling setVisible(false) on the platform window, so we don't
have the needed information anymore to check if the window we are
hiding was the focus-window. To work around this we use the native
state instead, which should still be valid if the window was the
focus window.

Change-Id: I98057e6393411471f03668e3e5ce544f6b49c01d
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@theqtcompany.com>
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@theqtcompany.com>