EGLFS: Make Mali GPU backend more generic
This provides both build and runtime fixes to make the Mali GPU backend
able to be used in more places.

The Mali backend was always trying to resize the frambuffer driver to
twice the default height.  In the case of the Mali T628 in the ODROID
XU3 it is an error to set the virtual height of the framebuffer device
to anything greater than the physical height.

Change-Id: I25a64e411f44d2aeb1cb4376183d9ea547c10161
Reviewed-by: Laszlo Agocs <laszlo.agocs@theqtcompany.com>