tst_QDBusAbstractAdaptor: fix reception of signals from P2P connection
P2P connections don't have senders and receivers, so asking
QDBusConnection to connect to a signal with a sender was a mistake
(added in 5368e44a86a4e0d4582ff5268986ea8bd0fa64ca). Due to an internal
bug, this never presented itself -- double fault.

Fix the connection so that we don't get unit test failures when the bug
is solved.

Change-Id: I9a75ad8521ae4e5cbbe5ffff13d1a78b7dea6d07
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>